+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- TestHelpers: addAndCommitVersions: appendToPubspec: can only append one time
+
 ## [2.5.5] - 2024-04-08
 
 ### Added
@@ -48,6 +54,7 @@
 - `HeadMessage`: Allow to return head message also when not everything is committed.
 - `modified-files`
 
+[Unreleased]: https://github.com/inlavigo/gg_git/compare/2.5.5...HEAD
 [2.5.5]: https://github.com/inlavigo/gg_git/compare/2.5.4...2.5.5
 [2.5.4]: https://github.com/inlavigo/gg_git/compare/2.5.3...2.5.4
 [2.5.3]: https://github.com/inlavigo/gg_git/compare/2.5.2...2.5.3